fix parquet benches
