+++ amadeus-parquet/src/internal/encodings/encoding.rs
@@ -998,13 +998,15 @@ impl Encoder<FixedLenByteArrayType> for DeltaByteArrayEncoder<FixedLenByteArrayT
 
 #[cfg(test)]
 mod tests {
-	use super::*;
-
 	use std::rc::Rc;
+	use test::Bencher;
 
+	use super::*;
 	use crate::internal::{
 		decoding::{get_decoder, Decoder, DictDecoder, PlainDecoder}, schema::types::{ColumnDescPtr, ColumnDescriptor, ColumnPath, Type as SchemaType}, util::{
-			memory::MemTracker, test_common::{random_bytes, RandGen}
+			memory::MemTracker, test_common::{
+				benches::{col_desc, gen_10, gen_100, gen_1000, gen_test_strs}, random_bytes, RandGen
+			}
 		}
 	};
 
@@ -1375,4 +1377,111 @@ mod tests {
 	fn create_test_dict_decoder<T: DataType>() -> DictDecoder<T> {
 		DictDecoder::<T>::new()
 	}
+
+	// Benches
+
+	macro_rules! plain {
+		($fname:ident, $batch_size:expr, $ty:ident, $pty:expr, $gen_data_fn:expr) => {
+			#[bench]
+			fn $fname(bench: &mut Bencher) {
+				let mem_tracker = Rc::new(MemTracker::new());
+				let encoder =
+					PlainEncoder::<$ty>::new(Rc::new(col_desc(0, $pty)), mem_tracker, vec![]);
+				let (bytes, values) = $gen_data_fn($batch_size);
+				bench_encoding(bench, bytes, values, Box::new(encoder));
+			}
+		};
+	}
+
+	macro_rules! dict {
+		($fname:ident, $batch_size:expr, $ty:ident, $pty:expr, $gen_data_fn:expr) => {
+			#[bench]
+			fn $fname(bench: &mut Bencher) {
+				let mem_tracker = Rc::new(MemTracker::new());
+				let encoder = DictEncoder::<$ty>::new(Rc::new(col_desc(0, $pty)), mem_tracker);
+				let (bytes, values) = $gen_data_fn($batch_size);
+				bench_encoding(bench, bytes, values, Box::new(encoder));
+			}
+		};
+	}
+
+	macro_rules! delta_bit_pack {
+		($fname:ident, $batch_size:expr, $ty:ident, $gen_data_fn:expr) => {
+			#[bench]
+			fn $fname(bench: &mut Bencher) {
+				let encoder = DeltaBitPackEncoder::<$ty>::new();
+				let (bytes, values) = $gen_data_fn($batch_size);
+				bench_encoding(bench, bytes, values, Box::new(encoder));
+			}
+		};
+	}
+
+	fn bench_encoding<T: DataType>(
+		bench: &mut Bencher, bytes: usize, values: Vec<T::Type>, mut encoder: Box<dyn Encoder<T>>,
+	) {
+		bench.bytes = bytes as u64;
+		bench.iter(|| {
+			encoder.put(&values[..]).expect("put() should be OK");
+			let _ = encoder.flush_buffer().expect("flush_buffer() should be OK");
+		})
+	}
+
+	plain!(plain_i32_1k_10, 1024, Int32Type, Type::Int32, gen_10);
+	plain!(plain_i32_1k_100, 1024, Int32Type, Type::Int32, gen_100);
+	plain!(plain_i32_1k_1000, 1024, Int32Type, Type::Int32, gen_1000);
+	plain!(plain_i32_1m_10, 1024 * 1024, Int32Type, Type::Int32, gen_10);
+	plain!(
+		plain_i32_1m_100,
+		1024 * 1024,
+		Int32Type,
+		Type::Int32,
+		gen_100
+	);
+	plain!(
+		plain_i32_1m_1000,
+		1024 * 1024,
+		Int32Type,
+		Type::Int32,
+		gen_1000
+	);
+	plain!(
+		plain_str_1m,
+		1024 * 1024,
+		ByteArrayType,
+		Type::ByteArray,
+		gen_test_strs
+	);
+
+	dict!(dict_i32_1k_10, 1024, Int32Type, Type::Int32, gen_10);
+	dict!(dict_i32_1k_100, 1024, Int32Type, Type::Int32, gen_100);
+	dict!(dict_i32_1k_1000, 1024, Int32Type, Type::Int32, gen_1000);
+	dict!(dict_i32_1m_10, 1024 * 1024, Int32Type, Type::Int32, gen_10);
+	dict!(
+		dict_i32_1m_100,
+		1024 * 1024,
+		Int32Type,
+		Type::Int32,
+		gen_100
+	);
+	dict!(
+		dict_i32_1m_1000,
+		1024 * 1024,
+		Int32Type,
+		Type::Int32,
+		gen_1000
+	);
+	plain!(
+		dict_str_1m,
+		1024 * 1024,
+		ByteArrayType,
+		Type::ByteArray,
+		gen_test_strs
+	);
+
+	delta_bit_pack!(delta_bit_pack_i32_1k_10, 1024, Int32Type, gen_10);
+	delta_bit_pack!(delta_bit_pack_i32_1k_100, 1024, Int32Type, gen_100);
+	delta_bit_pack!(delta_bit_pack_i32_1k_1000, 1024, Int32Type, gen_1000);
+	delta_bit_pack!(delta_bit_pack_i32_1m_10, 1024 * 1024, Int32Type, gen_10);
+	delta_bit_pack!(delta_bit_pack_i32_1m_100, 1024 * 1024, Int32Type, gen_100);
+	delta_bit_pack!(delta_bit_pack_i32_1m_1000, 1024 * 1024, Int32Type, gen_1000);
 }
