+++ amadeus-parquet/src/internal/util/test_common/mod.rs
@@ -15,6 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
+pub mod benches;
 pub mod file_util;
 pub mod page_util;
 pub mod rand_gen;
